+++ pkg/webui/console/store/selectors/logout.js
@@ -15,7 +15,7 @@
 import { createFetchingSelector } from '@ttn-lw/lib/store/selectors/fetching'
 import { createErrorSelector } from '@ttn-lw/lib/store/selectors/error'
 
-import { GET_USER_ME_BASE, LOGOUT_BASE } from '@console/store/actions/user'
+import { GET_USER_ME_BASE, LOGOUT_BASE } from '@console/store/actions/logout'
 
 const selectUserStore = state => state.user