+++ pkg/webui/console/views/user-api-key-edit/connect.js
@@ -22,7 +22,7 @@ import withRequest from '@ttn-lw/lib/components/with-request'
 import { getApiKey } from '@console/store/actions/api-keys'
 import { getUsersRightsList } from '@console/store/actions/users'
 
-import { selectUserId } from '@console/store/selectors/user'
+import { selectUserId } from '@console/store/selectors/logout'
 import {
   selectUserRights,
   selectUserPseudoRights,